コードを整理
いくつかのコードを1行にまとめました。

+++ memoapp.rb
@@ -4,36 +4,22 @@ require 'sinatra/reloader'
 require 'json'
 require 'securerandom'
 
-
-
 get "/memos" do
     @title = "メモアプリ"
-    unless FileTest.exist?("db/memo.json")
-      File.open("db/memo.json", "w")
-    end
-    @memo_data = if File.read("db/memo.json",1).nil?
-      nil
-    else
-      JSON.parse(File.read("db/memo.json"), symbolize_names: true)
-    end
+    File.open("db/memo.json", "w") unless FileTest.exist?("db/memo.json")
+    @memo_data = File.read("db/memo.json",1).nil? ? nil : JSON.parse(File.read("db/memo.json"), symbolize_names: true)
     erb :top
 end
 
 post "/memos" do
-  if File.read("db/memo.json",1).nil?
-    memos = []
-  else
-    memos = JSON.parse(File.read("db/memo.json"), symbolize_names: true)
-  end
+  File.read("db/memo.json",1).nil? ? memos = [] : memos = JSON.parse(File.read("db/memo.json"), symbolize_names: true)
   memo = {
     title: params[:title],
     contents: params[:contents],
     id: SecureRandom.uuid
   }
   memos << memo
-  File.open("db/memo.json", "w") do |file|
-    JSON.dump(memos, file)
-  end
+  File.open("db/memo.json", "w") { |file| JSON.dump(memos, file) }
   redirect "/memos"
 end
 
@@ -42,8 +28,10 @@ get "/memos/new" do
 end
 
 get "/memos/:id" do
-  memo_data = JSON.parse(File.read("db/memo.json"), symbolize_names: true)
-  @title = memo_data.find { |file| file[:id] == params[:id] }[:title]
-  @contents = memo_data.find { |file| file[:id] == params[:id] }[:contents]
+  memo_data = JSON.parse(File.read("db/memo.json"), symbolize_names: true).find { |file| file[:id] == params[:id] }
+  @title = memo_data[:title]
+  @contents = memo_data[:contents]
   erb :show
 end
+
+# SinataraアプリのGitについて調べる
